Improve event triggers for crates.io install test
This moves the `installation` test, which tests if installation
works from crates.io, out of the main `ci.yml` workflow and to a
new workflow `install.yml` that runs after a new release is made
(rather than when commits are pushed or pull requests are opened or
updated), as well as once a week and if it is manually triggered.

Multiple events occur in connection with a release. This includes
the creation and publication of the GitHub release. The could
trigger on an explicitly release-related event, but it instead
triggers on the `push` event, for tags only, where the tag starts
with `v`. All these events take place after the crate has been
uploaded to crates.io for the release, so installation should
succeed.

The reason for running when the version tag is pushed is that,
since `cargo-smart-release` publishes itself, and it creates and
pushes the version tag before it creates the GitHub release, this
will allow it to begin running sooner and thus catch and notify
developers of any installation problems a bit earlier:

https://github.com/GitoxideLabs/cargo-smart-release/blob/d59c889b5cd931ac22c27e02ef2af063834f6b11/src/command/release/mod.rs#L445-L456

But it would also be fine to trigger on a GitHub-release related
event, and it could be changed to that if the need arises.

For scheduled runs, this runs once a week and is triggered on the
same day and at the same time as the `rust-next.yml` workflow. This
is so that their result can be compared, mainly with the idea that
if `rust-next.yml` fails unexpectedly, the `install.yml` run may
help indicate whether recent unpublished changes in
`cargo-smart-release` are plausibly the cause.

Fixes #56