Downgrade `pulldown-cmark` to 0.9.6; bump some others
This commit downgrades `pulldown-cmark` from 0.13.0 to 0.9.6 in
`Cargo.toml` and `Cargo.lock`.

0.9.6 is the latest 0.9.* version of `pulldown-cmark`. Going to
0.10.* or higher would require code changes to be able to build,
and further changes to preserve correct functionality.

It is currently uncertain what those further changes must be. For
details, see the discussion in:
https://github.com/GitoxideLabs/cargo-smart-release/pull/54

This commit also upgrades `jiff` from 0.2.8 to 0.2.9, and upgrades
`signal-hook-registry` (in `Cargo.lock` only) from 1.4.2 to 1.4.5.