Replace another use of `.then()` with `if`-`then`
Clippy has begun to disapprove of such uses that can be expressed
no less complicatedly with `if`-`then`, but while it was able to
detect this one like the others, it was not able to auto-fix it.

A direct refactoring might use `Default::default()`, but
`"".into()` seems slightly more readable here for showing the
effect on the message that gets built into the log.