Don't include `pulldown-cmark` in grouped version updates
This is an attempt to work around the problem seen in #85 and
discussed in comments there, where even though we prohibit version
updates to `pulldown-cmark` already (see `dependendabot.yml`, #54,
and #59), Dependabot has recently begun to attempt them anyway.

This may not prevent Dependabot from attempting to update the crate
to a version it's not supposed to, but it should cause any such
attempts to be done in their own PRs, separate from the main
grouped PR where other updates are done.